requirements.txt
django
djangorestframework
django-cors-headers
Pillow
#swagger
drf-yasg
openpyxl
dj-database-url